requirements.txt
Mako==1.1.1
Markdown==3.1.1
numpy==1.16.4
matplotlib
hydra-core==0.11.3
opencv
torch 
torchvision 
torchaudio
